requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pyjwt
passlib[bcrypt]
pydantic[email]
bcrypt<4.0.0